Refactor and enhance various components across the codebase
- Added `#[allow(clippy::too_many_arguments)]` annotations to several functions to suppress warnings related to argument count.
- Updated `TOOL_SCHEMA_VERSION` to 16, reflecting new features such as optional width/height for shape generators and anchor points for clip transformations.
- Modified assertions in tests to improve clarity and accuracy.
- Refined JSON schema definitions in snapshots to include new properties for clip transformations.
- Improved function signatures and internal logic for better readability and performance.

These changes aim to enhance code quality and maintainability while introducing new features and improving existing functionality.

+++ crates/cutlass-compositor/tests/yuv_convert.rs
@@ -17,9 +17,9 @@ fn solid_yuv420p(width: u32, height: u32, y: u8, u: u8, v: u8) -> Yuv420pLayer {
         vec![y; w * h],
         width,
         vec![u; (w / 2) * (h / 2)],
-        (width / 2) as u32,
+        width / 2,
         vec![v; (w / 2) * (h / 2)],
-        (width / 2) as u32,
+        width / 2,
     )
 }
